Add tests
So that we can easily check that we have not broken anything.

The tests will create a database each time they run but that's a small
cost as it should only happen once.

The credentials are hard coded to default ones as I don't want to spend
time configuring it.